scoreboard class extracted
